[refactor] update category constant and RepeatDayType
