+++ feature/src/main/java/com/begin_a_gain/feature/match/common/MatchCategory.kt
@@ -17,17 +17,11 @@ import androidx.compose.ui.layout.Layout
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
+import com.begin_a_gain.feature.match.create_match.util.constant.categories
 import com.begin_a_gain.library.design.component.text.OText
 import com.begin_a_gain.library.design.theme.ColorToken
 import com.begin_a_gain.library.design.theme.ColorToken.Companion.color
 import com.begin_a_gain.library.design.theme.OTextStyle
-
-val categories: List<Pair<String, String>> = listOf(
-    Pair("\u200D\uD83D\uDCAA\uFE0F", "운동"), Pair("\uD83C\uDFC3\u200D", "건강"),
-    Pair("\uD83D\uDCD5", "독서"), Pair("\uD83C\uDF08", "취미"), Pair("\uD83E\uDEE7", "생활"),
-    Pair("\uD83D\uDCDD", "공부"), Pair("\uD83D\uDD01", "연습"), Pair("\uD83D\uDDDE", "시사/교양"),
-    Pair("\uD83E\uDD57", "다이어트"), Pair("\uD83D\uDCDA", "자기계발"), Pair("\uD83C\uDFB5", "악기")
-)
 
 @Preview(showSystemUi = true)
 @Composable
