+++ feature/src/main/java/com/begin_a_gain/feature/match/create_match/util/constant/Category.kt
@@ -1,0 +1,15 @@
+package com.begin_a_gain.feature.match.create_match.util.constant
+
+val categories: List<Pair<String, String>> = listOf(
+    Pair("\u200D\uD83D\uDCAA\uFE0F", "운동"),
+    Pair("\uD83C\uDFC3\u200D", "건강"),
+    Pair("\uD83D\uDCD5", "독서"),
+    Pair("\uD83C\uDF08", "취미"),
+    Pair("\uD83E\uDEE7", "생활"),
+    Pair("\uD83D\uDCDD", "공부"),
+    Pair("\uD83D\uDD01", "연습"),
+    Pair("\uD83D\uDDDE", "시사/교양"),
+    Pair("\uD83E\uDD57", "다이어트"),
+    Pair("\uD83D\uDCDA", "자기계발"),
+    Pair("\uD83C\uDFB5", "악기")
+)
